adds bar and pie charts
