+++ app/db.py
@@ -1,36 +1,47 @@
-"""Database functions"""
 import os
 
+import pandas as pd
+import psycopg2
 from dotenv import load_dotenv
-from fastapi import APIRouter, Depends
-import sqlalchemy
 
 
-router = APIRouter()
+load_dotenv()
+db_url = os.getenv('DB_URL')
 
 
-async def get_db() -> sqlalchemy.engine.base.Connection:
-    """ Get a SQLAlchemy database connection.
-    Uses this environment variable if it exists:  
-    DATABASE_URL=dialect://user:password@host/dbname
-    Otherwise uses a SQLite database for initial local development.
-    """
-    load_dotenv()
-    database_url = os.getenv('DB_URL', default='sqlite:///temporary.db')
-    engine = sqlalchemy.create_engine(database_url)
-    connection = engine.connect()
-    try:
-        yield connection
-    finally:
-        connection.close()
+def db_action(sql_action: str):
+    """ DB Setter - Performs a DB action returns None """
+    conn = psycopg2.connect(db_url)
+    curs = conn.cursor()
+    curs.execute(sql_action)
+    conn.commit()
+    curs.close()
+    conn.close()
 
 
-@router.get('/info')
-async def get_url(connection=Depends(get_db)):
-    """ Verify we can connect to the database,
-    and return the database URL in this format:
-    dialect://user:password@host/dbname
-    The password will be hidden with ***
-    """
-    url_without_password = repr(connection.engine.url)
-    return {'database_url': url_without_password}
+def db_query(sql_query: str) -> list:
+    """ DB Getter - Returns query results as a list """
+    conn = psycopg2.connect(db_url)
+    curs = conn.cursor()
+    curs.execute(sql_query)
+    results = curs.fetchall()
+    curs.close()
+    conn.close()
+    return results
+
+
+def get_df() -> pd.DataFrame:
+    """ Gets a DataFrame representation of the DB """
+    conn = psycopg2.connect(db_url)
+    curs = conn.cursor()
+    curs.execute("""SELECT clubname AS club, activityname AS activity, reactionint AS sentiment
+    FROM memberreactions
+    JOIN reactions ON memberreactions.reactionid = reactions.reactionid
+    JOIN activities ON memberreactions.activityid = activities.activityid
+    JOIN clubs ON memberreactions.clubid = clubs.clubid;""")
+    cols = [k[0] for k in curs.description]
+    rows = curs.fetchall()
+    df = pd.DataFrame(rows, columns=cols)
+    curs.close()
+    conn.close()
+    return df
