--- app/ml.py
@@ -1,5 +1,0 @@
-"""Machine learning functions"""
-from fastapi import APIRouter
-
-
-router = APIRouter()
