+++ requirements.txt
@@ -4,5 +4,4 @@ pandas~=1.3.0
 plotly~=5.1.0
 psycopg2-binary~=2.9.1
 python-dotenv~=0.18.0
-SQLAlchemy~=1.4.20
 uvicorn[standard]~=0.14.0
